requirements.txt
langchain
langchain_community
openai
streamlit 
newsapi-python
python-dotenv
